Refactor partner waitlist
